Speichern und Abrufen in Datenbank möglich

+++ whattowear/src/main/java/com/webtech/whattowear/repository/ClothesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClothesRepository extends CrudRepository <Clothes,Integer> {
+public interface ClothesRepository extends CrudRepository <Clothes,Long> {
 }
